Add authentication mechanism "passport-remember-me"
Add the possibility to be remembered (now, up to 7 days) without
reinserting the password. Using the auth mechanism from passport one can
just not insert password and have a cookie remember him

Refactor code, for adding a file token.js that handles the temporary
token (TODO: add Redis) and the authentication of the module that
requests an API token.

Modify registration to keep it simple, just for the sake of the demo